requirements.txt
pandas==2.1.1
numpy==1.23.5
apache-airflow==2.6.3  # Si Airflow est utilisé pour l'orchestration
